peso y altura implementados + dataset shuffle

+++ MutritionIA/Usuario.py
@@ -3,5 +3,5 @@ from pydantic import BaseModel
 class Usuario(BaseModel):
     name : str
     peso : int
-    altura : float
+    altura : int
     alergias : List
